requirements.txt
streamlit
openai>=1.0.0
kerykeion
swisseph
geopy
reportlab
pillow
timezonefinder==6.2.0